api_datos/pyproject.toml
[project]
name = "rugpull-api"
version = "2.0.0"
description = "Rug Pull Batch API — DuckDB + Google Drive + FastAPI"
requires-python = ">=3.11,<3.14"   
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "duckdb==1.1.3",
    "requests==2.32.3",
    "pydantic==2.11.1",           
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "python-dotenv==1.0.1",
]

[dependency-groups]               
dev = [
    "httpx>=0.27.0",              
]